Enhancing Functionality
Added a docstring to the script function to explain what the script does, as well as an output option to allow the user to choose where the output will be written to (default is console) and a filename option to specify the filename when writing to a file (default is output.txt). Additionally, an if name == 'main': block was added to ensure that the script() function is only called if the script is run directly, and not if it is imported as a module.